# docs/public/openapi.json
{
  "components": {
    "schemas": {
...
      "CanData": {
        "description": "Данные CAN-шины.",
        "properties": {
          "alarm_flags": {
            "default": 0,
            "title": "Alarm Flags",
            "type": "integer"
          },
          "engine_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Моточасы, ч",
            "title": "Engine Hours"
          },
          "engine_rpm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Rpm"
          },
          "engine_temp_c": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Temp C"
          },
          "odometer_km": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Полный пробег, км",
            "title": "Odometer Km"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Скорость ТС, км/ч",
            "title": "Speed Kmh"
          }
        },
        "title": "CanData",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NavData": {
        "description": "Навигационные данные telemetry-события.",
        "properties": {
          "alert_flag": {
            "default": false,
            "title": "Alert Flag",
            "type": "boolean"
          },
          "altitude_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude M"
          },
          "battery_voltage_mv": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Напряжение батареи, мВ (1 единица NDTP = 20 мВ)",
            "title": "Battery Voltage Mv"
          },
          "coordinates_valid": {
            "default": true,
            "title": "Coordinates Valid",
            "type": "boolean"
          },
          "course": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Курс, 0..360 градусов",
            "title": "Course"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "satellites": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Satellites"
          },
          "sos_flag": {
            "default": false,
            "title": "Sos Flag",
            "type": "boolean"
          },
          "speed_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Средняя скорость, км/ч",
            "title": "Speed Avg"
          },
          "speed_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Максимальная скорость, км/ч",
            "title": "Speed Max"
          },
          "timestamp": {
            "description": "Unix time устройства, секунды",
            "title": "Timestamp",
            "type": "integer"
          },
          "track_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Пройденный путь, м",
            "title": "Track M"
          }
        },
        "required": [
          "timestamp",
          "latitude",
          "longitude"
        ],
        "title": "NavData",
        "type": "object"
      },
...
      "ReplayFleet": {
        "properties": {
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "source": {
            "default": "historical_csv_test",
            "title": "Source",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/ReplayFleetVehicle"
            },
            "title": "Vehicles",
            "type": "array"
          }
        },
        "required": [
          "start_at",
          "end_at",
          "vehicles"
        ],
        "title": "ReplayFleet",
        "type": "object"
      },
      "ReplayFleetVehicle": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/ReplayPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/ReplayRun"
            },
            "title": "Runs",
            "type": "array"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/ReplayStop"
            },
            "title": "Stops",
            "type": "array"
          },
          "telemetry": {
            "items": {
              "$ref": "#/components/schemas/ReplayTelemetry"
            },
            "title": "Telemetry",
            "type": "array"
          },
          "tr_id": {
            "title": "Tr Id",
            "type": "integer"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "integer"
          }
        },
        "required": [
          "tr_id",
          "unit_id",
          "telemetry",
          "stops",
          "runs",
          "points"
        ],
        "title": "ReplayFleetVehicle",
        "type": "object"
      },
      "ReplayOutcome": {
        "properties": {
          "actual_at": {
            "format": "date-time",
            "title": "Actual At",
            "type": "string"
          },
          "actual_delay_s": {
            "title": "Actual Delay S",
            "type": "number"
          },
          "sample_id": {
            "title": "Sample Id",
            "type": "string"
          }
        },
        "required": [
          "sample_id",
          "actual_delay_s",
          "actual_at"
        ],
        "title": "ReplayOutcome",
        "type": "object"
      },
      "ReplayPoint": {
        "properties": {
          "T": {
            "format": "date-time",
            "title": "T",
            "type": "string"
          },
          "actual_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual At"
          },
          "actual_delay_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Delay S"
          },
          "cur_dev_s": {
            "title": "Cur Dev S",
            "type": "number"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "sample_id": {
            "title": "Sample Id",
            "type": "string"
          },
          "target_stop_id": {
            "title": "Target Stop Id",
            "type": "integer"
          },
          "target_time_begin": {
            "format": "date-time",
            "title": "Target Time Begin",
            "type": "string"
          }
        },
        "required": [
          "sample_id",
          "T",
          "target_stop_id",
          "target_time_begin",
          "cur_dev_s",
          "lat",
          "lon"
        ],
        "title": "ReplayPoint",
        "type": "object"
      },
      "ReplayPrediction": {
        "properties": {
          "baseline_delay_s": {
            "title": "Baseline Delay S",
            "type": "number"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "predicted_delay_s": {
            "title": "Predicted Delay S",
            "type": "number"
          },
          "sample_id": {
            "title": "Sample Id",
            "type": "string"
          }
        },
        "required": [
          "sample_id",
          "predicted_delay_s",
          "baseline_delay_s",
          "model_version"
        ],
        "title": "ReplayPrediction",
        "type": "object"
      },
      "ReplayRun": {
        "properties": {
          "confirmed_stops": {
            "title": "Confirmed Stops",
            "type": "integer"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "stop_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Stop Ids",
            "type": "array"
          },
          "tr_id": {
            "title": "Tr Id",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "run_id",
          "tr_id",
          "start_at",
          "end_at",
          "stop_ids",
          "confirmed_stops",
          "valid"
        ],
        "title": "ReplayRun",
        "type": "object"
      },
      "ReplayScenario": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/ReplayPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "source": {
            "default": "historical_csv_test",
            "title": "Source",
            "type": "string"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/ReplayStop"
            },
            "title": "Stops",
            "type": "array"
          },
          "telemetry": {
            "items": {
              "$ref": "#/components/schemas/ReplayTelemetry"
            },
            "title": "Telemetry",
            "type": "array"
          },
          "vehicle": {
            "$ref": "#/components/schemas/ReplayVehicle"
          }
        },
        "required": [
          "vehicle",
          "telemetry",
          "stops",
          "points"
        ],
        "title": "ReplayScenario",
        "type": "object"
      },
      "ReplayStop": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "gps_confirmed": {
            "default": false,
            "title": "Gps Confirmed",
            "type": "boolean"
          },
          "gps_confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Confirmed At"
          },
          "gps_distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Distance M"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "place_id": {
            "default": "",
            "title": "Place Id",
            "type": "string"
          },
          "planned_at": {
            "format": "date-time",
            "title": "Planned At",
            "type": "string"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          }
        },
        "required": [
          "stop_id",
          "planned_at",
          "lat",
          "lon"
        ],
        "title": "ReplayStop",
        "type": "object"
      },
      "ReplayStreamStatus": {
        "properties": {
          "sent_packets": {
            "title": "Sent Packets",
            "type": "integer"
          },
          "source_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source At"
          },
          "speed_multiplier": {
            "title": "Speed Multiplier",
            "type": "number"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "time_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Mode"
          },
          "total_packets": {
            "title": "Total Packets",
            "type": "integer"
          }
        },
        "required": [
          "state",
          "source_at",
          "time_mode",
          "speed_multiplier",
          "sent_packets",
          "total_packets"
        ],
        "title": "ReplayStreamStatus",
        "type": "object"
      },
      "ReplayTelemetry": {
        "properties": {
          "available_at": {
            "format": "date-time",
            "title": "Available At",
            "type": "string"
          },
          "event_time": {
            "format": "date-time",
            "title": "Event Time",
            "type": "string"
          },
          "heading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          }
        },
        "required": [
          "available_at",
          "event_time",
          "lat",
          "lon",
          "speed"
        ],
        "title": "ReplayTelemetry",
        "type": "object"
      },
      "ReplayVehicle": {
        "properties": {
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "samples": {
            "title": "Samples",
            "type": "integer"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "tr_id": {
            "title": "Tr Id",
            "type": "integer"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "integer"
          }
        },
        "required": [
          "tr_id",
          "unit_id",
          "samples",
          "start_at",
          "end_at"
        ],
        "title": "ReplayVehicle",
        "type": "object"
      },
...
      "TelemetryEvent": {
        "description": "Нормализованная телеметрия одного устройства.\n\nЕдиный внутренний формат: остальное приложение работает только с этой\nмоделью и не знает ничего о протоколе NDTP.",
        "properties": {
          "can": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanData"
              },
              {
                "type": "null"
              }
            ]
          },
          "nav": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NavData"
              },
              {
                "type": "null"
              }
            ]
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "unit_id": {
            "description": "ID терминала (peerAddress NDTP)",
            "title": "Unit Id",
            "type": "integer"
          }
        },
        "required": [
          "unit_id",
          "received_at"
        ],
        "title": "TelemetryEvent",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Backend API системы раннего прогнозирования задержек наземного транспорта.\n\n1. Получите токен через `POST /api/v1/auth/token`.\n2. В Swagger UI нажмите **Authorize** и введите учётные данные.\n3. Health check остаётся публичным; telemetry и prediction API защищены.",
    "title": "Transport Delay Predictor API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/replay/fleet": {
      "get": {
        "operationId": "get_replay_fleet_api_v1_replay_fleet_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayFleet"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Общая январская шкала, GPS и проверенные рейсы",
        "tags": [
          "Historical replay"
        ]
      }
    },
    "/api/v1/replay/outcomes": {
      "get": {
        "operationId": "get_replay_outcomes_api_v1_replay_outcomes_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "As Of",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReplayOutcome"
                  },
                  "title": "Response Get Replay Outcomes Api V1 Replay Outcomes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Факты прибытия, уже наступившие к моменту replay",
        "tags": [
          "Historical replay"
        ]
      }
    },
    "/api/v1/replay/predict/{sample_id}": {
      "post": {
        "operationId": "predict_replay_point_api_v1_replay_predict__sample_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayPrediction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Прогноз по январским данным, доступным к моменту T",
        "tags": [
          "Historical replay"
        ]
      }
    },
    "/api/v1/replay/stream-status": {
      "get": {
        "operationId": "get_replay_stream_status_api_v1_replay_stream_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayStreamStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Исходные часы активного NDTP-воспроизведения",
        "tags": [
          "Historical replay"
        ]
      }
    },
    "/api/v1/replay/vehicles": {
      "get": {
        "operationId": "list_replay_vehicles_api_v1_replay_vehicles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReplayVehicle"
                  },
                  "title": "Response List Replay Vehicles Api V1 Replay Vehicles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Транспорт с известным фактом в январской записи",
        "tags": [
          "Historical replay"
        ]
      }
    },
    "/api/v1/replay/vehicles/{tr_id}": {
      "get": {
        "operationId": "get_replay_scenario_api_v1_replay_vehicles__tr_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tr_id",
            "required": true,
            "schema": {
              "title": "Tr Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayScenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Телеметрия, план и точки прогноза для исторического прогона",
        "tags": [
          "Historical replay"
        ]
      }
    },
    "/api/v1/vehicles": {
      "get": {
        "operationId": "list_vehicles_api_v1_vehicles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TelemetryEvent"
                  },
                  "title": "Response List Vehicles Api V1 Vehicles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Список ТС с последней телеметрией",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{unit_id}": {
      "get": {
        "operationId": "get_vehicle_api_v1_vehicles__unit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "title": "Unit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Последняя телеметрия ТС",
        "tags": [
          "Vehicles"
        ]
      }
    },
...
  },
...
}